added react-icon library, modified App component to grab contactList data from third part server using useEffect & also added styles to the searchField component
