--- src/App.js
@@ -1,14 +1,0 @@
-
-import './App.css';
-
-const App=() =>{
-  return (
-    <div className="App">
-      <header className="App-header">
-        <h1>Contacts</h1>
-      </header>
-    </div>
-  );
-}
-
-export default App;
